+++ src/bandersnatch/tests/test_mirror.py
@@ -441,6 +441,107 @@ simple{0}index.html""".format(
 
 
 @pytest.mark.asyncio
+async def test_mirror_sync_package_download_mirror(
+    mirror: BandersnatchMirror,
+) -> None:
+    mirror.master.all_packages = mock.AsyncMock(return_value={"foo": 1})  # type: ignore
+    mirror.json_save = True
+    # Recall bootstrap so we have the json dirs
+    mirror._bootstrap()
+    # This download mirror URL works, forcing not to fallback
+    mirror.download_mirror = "https://pypi-mirror.example.com/pypi"
+    mirror.download_mirror_no_fallback = True
+    await mirror.synchronize()
+
+    assert """\
+json{0}foo
+last-modified
+packages{0}2.7{0}f{0}foo{0}foo.whl
+packages{0}any{0}f{0}foo{0}foo.zip
+pypi{0}foo{0}json
+simple{0}foo{0}index.html
+simple{0}index.html""".format(
+        sep
+    ) == utils.find(
+        mirror.webdir, dirs=False
+    )
+    assert (
+        open("web{0}simple{0}index.html".format(sep)).read()
+        == """\
+<!DOCTYPE html>
+<html>
+  <head>
+    <title>Simple Index</title>
+  </head>
+  <body>
+    <a href="foo/">foo</a><br/>
+  </body>
+</html>"""
+    )
+    assert open("status", "rb").read() == b"1"
+
+
+@pytest.mark.asyncio
+async def test_mirror_sync_package_download_mirror_fallback(
+    mirror: BandersnatchMirror,
+) -> None:
+    mirror.master.all_packages = mock.AsyncMock(return_value={"foo": 1})  # type: ignore
+    mirror.json_save = True
+    # Recall bootstrap so we have the json dirs
+    mirror._bootstrap()
+    # This download mirror URL does not work, should fallback to normal logic
+    mirror.download_mirror = "https://not-working.example.com/pypi"
+    await mirror.synchronize()
+
+    assert """\
+json{0}foo
+last-modified
+packages{0}2.7{0}f{0}foo{0}foo.whl
+packages{0}any{0}f{0}foo{0}foo.zip
+pypi{0}foo{0}json
+simple{0}foo{0}index.html
+simple{0}index.html""".format(
+        sep
+    ) == utils.find(
+        mirror.webdir, dirs=False
+    )
+    assert (
+        open("web{0}simple{0}index.html".format(sep)).read()
+        == """\
+<!DOCTYPE html>
+<html>
+  <head>
+    <title>Simple Index</title>
+  </head>
+  <body>
+    <a href="foo/">foo</a><br/>
+  </body>
+</html>"""
+    )
+
+
+@pytest.mark.asyncio
+async def test_mirror_sync_package_download_mirror_fails(
+    mirror: BandersnatchMirror,
+) -> None:
+    mirror.master.all_packages = mock.AsyncMock(return_value={"foo": 1})  # type: ignore
+    mirror.json_save = True
+    # Recall bootstrap so we have the json dirs
+    mirror._bootstrap()
+    # This download mirror URL does not work, forcing not to fallback
+    mirror.download_mirror = "https://not-working.example.com"
+    mirror.download_mirror_no_fallback = True
+
+    with mock.patch("bandersnatch.mirror.logger.exception") as mock_log:
+        await mirror.synchronize()
+        # 3 calls are:
+        # * Tried and failed to get /packages/any/f/foo/foo.zip
+        # * Tried and failed to get /packages/2.7/f/foo/foo.whl
+        # * Declared error on syncing package foo@1
+        assert mock_log.call_count == 3
+
+
+@pytest.mark.asyncio
 async def test_mirror_serial_current_no_sync_of_packages_and_index_page(
     mirror: BandersnatchMirror,
 ) -> None:
@@ -898,14 +999,30 @@ async def test_package_sync_does_not_touch_existing_local_file(
     assert old_stat.st_ctime == Path(pkg_file_path_str).stat().st_ctime
 
 
-def test_gen_data_requires_python(mirror: BandersnatchMirror) -> None:
+def test_gen_html_file_tags(mirror: BandersnatchMirror) -> None:
     fake_no_release: Dict[str, str] = {}
-    fake_release = {"requires_python": ">=3.6"}
 
-    assert mirror.gen_data_requires_python(fake_no_release) == ""
+    # only requires_python
+    fake_release_1 = {"requires_python": ">=3.6"}
+
+    # only data_yanked
+    fake_release_2 = {"yanked": True, "yanked_reason": "Broken release"}
+
+    # requires_python and data_yanked
+    fake_release_3 = {
+        "requires_python": ">=3.6",
+        "yanked": True,
+        "yanked_reason": "Broken release",
+    }
+
+    assert mirror.gen_html_file_tags(fake_no_release) == ""
     assert (
-        mirror.gen_data_requires_python(fake_release)
-        == ' data-requires-python="&gt;=3.6"'
+        mirror.gen_html_file_tags(fake_release_1) == ' data-requires-python="&gt;=3.6"'
+    )
+    assert mirror.gen_html_file_tags(fake_release_2) == ' data-yanked="Broken release"'
+    assert (
+        mirror.gen_html_file_tags(fake_release_3)
+        == ' data-requires-python="&gt;=3.6" data-yanked="Broken release"'
     )